requirements.txt
azure-ai-documentintelligence
azure-ai-inference
azure-ai-textanalytics
azure-common 
azure-core
azure-identity
Flask==3.0.3
matplotlib==3.5.1
opencv_contrib_python==4.10.0.84
pandas==2.2.2
Pillow==9.0.1
pyspellchecker==0.8.1
python-dotenv==1.0.1
Requests==2.32.3
transformers==4.42.3
